Move identities operations within their own package in go.server.
