+++ src/backend/interfaces/REST/go.server/operations/identities/identities.go
@@ -1,8 +1,14 @@
+/*
+ * // Copyleft (ɔ) 2018 The Caliopen contributors.
+ * // Use of this source code is governed by a GNU AFFERO GENERAL PUBLIC
+ * // license (AGPL) that can be found in the LICENSE file.
+ */
+
 // Copyleft (ɔ) 2017 The Caliopen contributors.
 // Use of this source code is governed by a GNU AFFERO GENERAL PUBLIC
 // license (AGPL) that can be found in the LICENSE file.
 
-package users
+package identities
 
 import (
 	"bytes"
